Calibration improvements and profiling
Added a semi-automated calibration development, where the temperatures
and calibration images are found based on directory names. Currently it
looks for 'cal' in the directory names and assumes the last two
characters in those names are the temperature at which the calibration
was taken.

Simple time profiling added.

Some formatting changes.

Rewrote experimentDirectory function in plif_tools to account for
alternate directory structure where the experiment images are in the
root directory instead of a sub.